+++ CHANGELOG.md
@@ -14,7 +14,8 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - Makes `device.isSandbox` more reliable.
 - Fixes the web restore alert not showing the "Yes" action button and "Cancel" incorrectly triggering the restore action.
 - Fixes a rare issue where a user's subscription could remain active after a refund, preventing paywalls from being shown.
-- Fixed trial eligibility for Stripe products.
+- Fixes trial eligibility for Stripe products.
+- Fixes an issue where `transaction_complete` could be missing transaction information when a crossgrade occurred while using a purchase controller.
 
 ## 4.14.0